- Angular - Training; + Shoping - List Project
