--- Angular-Training/src/app/assignment/assignment.component.scss
@@ -1,3 +1,0 @@
-.white {
-    color:white;
-}
