--- Angular-Training/src/app/server/server.component.html
@@ -1,5 +1,0 @@
-<p 
-    [ngStyle]="{backgroundColor: getColor()}"
-    [ngClass]="{online: serverStatus === 'online'}">
-    Server with ID {{ serverId }} is {{ getServerStatus() }}
-</p>
