--- Angular-Training/src/app/server/server.component.ts
@@ -1,27 +1,0 @@
-import {Component} from '@angular/core';
-
-@Component({
-  selector: 'app-server',
-  templateUrl: './server.component.html',
-  styles: [`
-        .online{
-            color: white;
-        }
-    `],
-})
-export class ServerComponent {
-  serverId: number = 10;
-  serverStatus: string = 'offline';
-
-  constructor() {
-    this.serverStatus = Math.random() > 0.5 ? 'online' : 'offline';
-  }
-
-  getServerStatus(): string {
-    return this.serverStatus;
-  }
-
-  getColor(): string {
-    return this.serverStatus === 'online' ? 'green' : 'red';
-  }
-}
