--- Angular-Training/src/app/servers/servers.component.html
@@ -1,23 +1,0 @@
-<label>Server Name</label>
-
-<!-- <input type="text" class="form-control" (input)="onUpdateServerName($event)"> -->
-
-<input type="text" class="form-control" [(ngModel)]="serverName">
-
-<!-- <p>{{serverName}}</p> -->
-
-<button class="btn btn-primary" [disabled]="!allowNewServer"
-        (click)="onCreateServer()">
-    Add Server
-</button>
-
-<!-- <p [innerText]="allowNewServer"></p> -->
-
-<!-- <p>
-    {{serverCreationStatus}}
-</p> -->
-
-<p *ngIf="serverCreated">Server was created, server name is {{serverName}}</p>
-
-
-<app-server *ngFor="let server of servers"></app-server>
