--- Angular-Training/src/app/servers/servers.component.ts
@@ -1,33 +1,0 @@
-import {Component, OnInit} from '@angular/core';
-
-@Component({
-  selector: 'app-servers',
-  templateUrl: './servers.component.html',
-  styleUrls: ['./servers.component.scss'],
-})
-export class ServersComponent implements OnInit {
-  allowNewServer: boolean = false;
-  serverCreationStatus: string = 'No Server was created!';
-  serverName: string = 'TestServer';
-  serverCreated: boolean = false;
-  servers: string[] = ['TestServer 1', 'TestServer 2'];
-
-  constructor() {
-    setTimeout(() => {
-      this.allowNewServer = true;
-    }, 2000 );
-  }
-
-  ngOnInit(): void {
-  }
-
-  onCreateServer(): void {
-    this.serverCreated = true;
-    this.servers.push(this.serverName);
-    this.serverCreationStatus = 'Server was created! Name is ' + this.serverName;
-  }
-
-  onUpdateServerName(event: Event): void {
-    this.serverName = (<HTMLInputElement>event.target).value;
-  }
-}
