+++ Shopping-App/README.md
@@ -1,4 +1,4 @@
-# AngularTraining
+# ShoppingApp
 
 This project was generated with [Angular CLI](https://github.com/angular/angular-cli) version 13.2.5.
