+++ Shopping-App/karma.conf.js
@@ -25,7 +25,7 @@ module.exports = function (config) {
       suppressAll: true // removes the duplicated traces
     },
     coverageReporter: {
-      dir: require('path').join(__dirname, './coverage/angular-training'),
+      dir: require('path').join(__dirname, './coverage/shopping-app'),
       subdir: '.',
       reporters: [
         { type: 'html' },
