+++ Shopping-App/package.json
@@ -1,5 +1,5 @@
 {
-  "name": "angular-training",
+  "name": "shopping-app",
   "version": "0.0.0",
   "scripts": {
     "ng": "ng",
@@ -29,6 +29,10 @@
     "@angular/compiler-cli": "~13.2.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
+    "@typescript-eslint/eslint-plugin": "^5.16.0",
+    "@typescript-eslint/parser": "^5.16.0",
+    "eslint": "^8.11.0",
+    "eslint-config-google": "^0.14.0",
     "jasmine-core": "~4.0.0",
     "karma": "~6.3.0",
     "karma-chrome-launcher": "~3.1.0",
